This commit removes report portal client from the main script and creates a module for it
Testing

It has been tested using the same command as the initial commit and successfully completed returning the correct amount of launches, suites and tests
